Added a footer and moved css files to a folder.
