+++ frontend/src/includes/Footer.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import '../styles/Footer.css';
+
+const Footer: React.FC = () => {
+    return (
+        <footer className="footer">
+            <div className="container">
+                <span className="text-muted">© Jacob och Linn</span>
+            </div>
+        </footer>
+    )
+    }
+export default Footer;
