+++ frontend/src/main.tsx
@@ -1,9 +1,8 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
-import './index.css'
+import './styles/index.css'
 import App from './App.tsx'
-import './Create.css'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
