+++ frontend/src/styles/Footer.css
@@ -1,0 +1,18 @@
+.footer {
+    background-color: #f8f9fa;
+    padding: 1rem 0;
+    margin-top: 2rem;
+    border-top: 1px solid #dee2e6;
+}
+
+.footer .container {
+    max-width: 1200px;
+    margin: 0 auto;
+    padding: 0 1rem;
+    text-align: center;
+}
+
+.footer .text-muted {
+    color: #6c757d;
+    font-size: 0.875rem;
+}
